Rework packaging to support SB2 and multiple target triples
Use a bootstrap approach aligned with Fedora packaging

The non-x86 builds produce just the stub binaries for SB2

The rust-std-static library packages now include their triple. On a
per-architecture basis the rust-std-static package is a virtual
Provide. This allows simple repackaging and host-cross-compiling in
non-x86 builds.

The patch applied to rust allows the default target to come from the
environment variable SB2_RUST_TARGET_TRIPLE.

The patch applied to cargo forces the target to be set for Host builds
otherwise SB2_RUST_TARGET_TRIPLE will make it default to arm

arm cc needs to find ld so ensure PATH points to /opt/cross/bin

The target .so files are not needed as we don't have target rustc
The *miri files don't seem to be built - leave a comment for now

aarch64 builds are disabled for now

Signed-off-by: David Greaves <david.greaves@jolla.com>