[rust] Bypass SB2 gated calls which deadlock. Contributes JB#50499
When rust's Command class does a spawn() call (at the rust level) it
invokes either libc::spawnvp() or libc::fork()/exec().
This is problematic because cargo is heavily threaded and fork/exec in
threads is known to be prone to deadlock due to futex sharing. (google
it!)

Whilst rust takes care to manage malloc, sb2 does not and calling sb2
between the fork and exec in a threaded application under SB2 will
lead to deadlock.  This patch introduces ENV triggered bypasses for
libc calls made in Command.spawn(). This includes, dup2, chdir and
close. This patch does not yet fix all hangs under sb2. (Eg the
close() call is invoked during the Drop trait of the shared pipe
IPC... there may be others hiding somewhere).

Don't strip when cross-building libs as it corrupts them

Signed-off-by: David Greaves <david.greaves@jolla.com>